+++ features/common.workflow-approvals.v1/configs/endpoints.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2020, WSO2 LLC. (https://www.wso2.com). All Rights Reserved.
+ * Copyright (c) 2020-2025, WSO2 LLC. (https://www.wso2.com).
  *
  * WSO2 LLC. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -16,8 +16,16 @@
  * under the License.
  */
 
-export * from "./api";
-export * from "./components";
-export * from "./configs";
-export * from "./models";
-export * from "./pages";
+import { ApprovalsResourceEndpointsInterface } from "../models";
+
+/**
+ * Get the resource endpoints for the Workflow approvals Management feature.
+ *
+ * @param serverHost - Server Host.
+ * @returns Resource endpoints for the Workflow approvals Management feature.
+ */
+export const getApprovalsResourceEndpoints = (serverHost: string): ApprovalsResourceEndpointsInterface => {
+    return {
+        approvals: `${ serverHost }/api/users/v1/me/approval-tasks`
+    };
+};