+++ features/common.workflow-approvals.v1/index.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2024, WSO2 LLC. (https://www.wso2.com).
+ * Copyright (c) 2020-2025, WSO2 LLC. (https://www.wso2.com).
  *
  * WSO2 LLC. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -16,25 +16,8 @@
  * under the License.
  */
 
-declare module "*.json" {
-    const value: any;
-
-    export default value;
-}
-
-declare module "*.svg" {
-    import React = require("react");
-
-    export const ReactComponent: React.FunctionComponent<React.SVGProps<SVGSVGElement>>;
-    const src: string;
-
-    export default src;
-}
-
-declare module "*.png" {
-    const content: string;
-
-    export default content;
-}
-
-declare module "*.md";
+export * from "./api";
+export * from "./components";
+export * from "./configs";
+export * from "./models";
+export * from "./pages";