+++ modules/i18n/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @wso2is/i18n
 
+## 2.26.2
+
+### Patch Changes
+
+- [#8592](https://github.com/wso2/identity-apps/pull/8592) [`3632b9dfe3fe2dbfd0a0d272b2bf1e1a43a6a523`](https://github.com/wso2/identity-apps/commit/3632b9dfe3fe2dbfd0a0d272b2bf1e1a43a6a523) Thanks [@Malith-19](https://github.com/Malith-19)! - Move the Approvals to common component.
+
 ## 2.26.1
 
 ### Patch Changes